feat: 회원가입 요청 DTO 약관동의항목 컬럼 추가 & 기존 테스트 오류 수정 UserControllerTest -> SignUpRequestTest로 DTO Validation 관련 테스트 로직 이동
