+++ src/main/resources/templates/user/login/signUpForm.html
@@ -105,7 +105,7 @@
                     <p>필수 항목: 이름, 이메일 주소, 비밀번호</p>
                     <br>
 
-                    <h6 style="color: #303038">2. 개인정보의 수집 및 이용 목적</h6>
+                    <h6 style="color: #303038">2. 개인정보수집 및 이용 목적</h6>
                     <p>제주도랑은 다음과 같은 목적으로 회원님의 개인정보를 수집하고 이용합니다.</p>
 
                     <p>
@@ -161,7 +161,7 @@
                 <div class="overflow-auto"
                      style="width:100%; height:100px; padding:15px; border: 1px solid #d0d0d0; border-radius: 10px; margin: 10px 0; background: #f8f8f8; color:#767678">
 
-                    <h5 style="color: #303038">제주도랑 선택 개인정보 수집 및 이용약관</h5><br>
+                    <h5 style="color: #303038">제주도랑 선택 개인정보 수집 및 이용 약관</h5><br>
 
                     <h6 style="color: #303038">1. 선택적으로 수집하는 개인정보 항목</h6>
                     <p>
@@ -172,7 +172,7 @@
                         주소 등 추가 정보<br>
                     </p>
 
-                    <h6 style="color: #303038">2. 선택적 개인정보의 수집 및 이용 목적</h6>
+                    <h6 style="color: #303038">2. 선택적 개인정보 수집 및 이용 목적</h6>
                     <p>
                         위의 선택적 개인정보는 아래와 같은 목적으로 수집 및 이용됩니다.<br><br>
